Update do_nothing transform so it converts image dtype to float (needed for pytorch)

+++ fnet/transforms.py
@@ -13,7 +13,7 @@ def normalize(img):
     return result
 
 def do_nothing(img):
-    return img
+    return img.astype(np.float)
 
 class Propper(object):
     """Padder + Cropper"""
